Refactor Theme to useContext
