+++ src/main.jsx
@@ -2,15 +2,15 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import { Theme, GlobalStyle } from "./components/Theme";
+import { ThemeWrapper, GlobalStyle } from "./components/Theme";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <Theme>
+    <ThemeWrapper>
       <GlobalStyle />
       <BrowserRouter>
         <App />
       </BrowserRouter>
-    </Theme>
+    </ThemeWrapper>
   </React.StrictMode>
 );
